refactor: extract submodules from combat, dungeon, and fishing logic
Split three large logic.rs files into focused submodules:
- combat: events.rs (types), damage.rs (enemy death), regen.rs (HP regen)
- dungeon: pathfinding.rs (room navigation), rewards.rs (treasure/XP)
- fishing: drops.rs (item drops), rank.rs (rank progression)

Deduplicates ~60-line enemy death handler that appeared twice in combat.
All public APIs preserved via re-exports. 2,973 tests pass, 92.7% coverage.
