+++ src/dungeon/mod.rs
@@ -4,8 +4,12 @@
 
 pub mod generation;
 pub mod logic;
+pub mod pathfinding;
+pub mod rewards;
 pub mod types;
 
 pub use generation::*;
 pub use logic::*;
+pub use pathfinding::*;
+pub use rewards::*;
 pub use types::*;
