+++ src/fishing/mod.rs
@@ -2,10 +2,13 @@
 
 #![allow(unused_imports)]
 
+pub mod drops;
 pub mod generation;
 pub mod logic;
+pub mod rank;
 pub mod types;
 
 pub use generation::*;
 pub use logic::*;
+pub use rank::*;
 pub use types::*;
