+++ tests/prestige_cycle_test.rs
@@ -230,9 +230,8 @@ fn test_cannot_prestige_when_ineligible() {
 fn test_combat_to_prestige_full_loop() {
     use quest::character::derived_stats::DerivedStats;
     use quest::character::prestige::{can_prestige, perform_prestige};
-    use quest::combat::logic::{
-        update_combat, CombatEvent, GodItemCombatBonuses, HavenCombatBonuses,
-    };
+    use quest::combat::logic::update_combat;
+    use quest::combat::{CombatEvent, GodItemCombatBonuses, HavenCombatBonuses};
     use quest::core::game_logic::spawn_enemy_if_needed;
     use quest::TICK_INTERVAL_MS;
